requirements.txt
# Core dependencies
flask==2.3.3
flask-cors==6.0.1
python-dotenv==1.1.0
openai==1.23.2
httpx==0.27.0
gunicorn==21.2.0

# AWS support (SecretsManager)
boto3==1.34.85
botocore==1.34.85

# (Optional ML dependencies)
scikit-learn==1.3.2
numpy==1.26.4
scipy==1.13.1
joblib==1.5.1
threadpoolctl==3.6.0
